MIPI:CODE&DTS:RzBoard: Add compatibility to support both 2-lanes and 4-lanes MIPI DSI
	Add a property 'dsi-lanes' in du node for rcar-du driver to load different parameters and clocks
	Add two resolution_param lists, corresponding to the needs of 2-lanes and 4-lanes DSI respectively
	Currently the display device as follow:
	HDMI (via IT6161 MIPI to HDMI chip) :    4-lanes MIPI DSI
	MIPI Screen (with 720*1280 resolutions): 2-lanes MIPI DSI

Because in order to support 2 lane MIPI DSI, the internal clock dividers need to be change.
Utill now the official driver only hard-code supports 4-lane MIPI DSI on RZ/V2L, and key Renesas
software resources do not have a schedule to make a change that can support both 2 lane and 4 lane.
So this commit is a patch fixed the incompatibility temporarily.

Signed-off-by: Nick <nick.cheng@avnet.com>